Hautobot

Description

Hautobot is a select-and-transform approach to HTML (and XML) templating. It is a highly inferior and highly incomplete implementation of the basic ideas in Christophe Grand's Enlive library for Clojure (https://github.com/cgrand/enlive).

Some Types

type Transform = Node -> [Node]

Transforms take a Node, and return none, one or many Nodes.

type Selector = [[Node -> Bool]]

A Selector is a list of lists of predicates. Each inner list represents a "descendant step". That is, each inner list must match a single node, and each subsequent inner list must match a node inside the previous. Examples:

 "a[href]"                       => [[hasTag "a", hasAttr "href"]]
 "#name.kls"                     => [[hasId "name", hasClass "kls"]]
 "section#name article.story h1" => [[hasTag "section", hasId "name"], 
                                     [hasTag "article", hasClass "story"],
                                     [hasTag "h1"]]

Predicates

hasId :: String -> Node -> Bool

Does the Node have the given id? (case-sensitive both on the attribute name id, and on the value.)

hasTag :: String -> Node -> Bool

Is the Node an element with the given tag name? (case-sensitive)

hasAttr :: String -> Node -> Bool

Does the Node have an attribute with the given name?

hasAttrWord

Arguments

:: String
attribute name

-> String
word to look for in attribute value

-> Node

-> Bool

Does the Node contain word (space-delimited) within attr?

hasClass :: String -> Node -> Bool

Does the Node have the given class?

Transformers

setAttr

Arguments

:: String
attribute name

-> String
attribute value

-> Transform

Set an attribute (overwriting any existing value)

setTag :: String -> Transform

Rename a tag

modifyAttr :: String -> (Maybe Text -> Maybe Text) -> Transform

Modify attribute with a function. Function is `Maybe (T.Text currentVal) -> Maybe (T.Text newVal)`. Returning Nothing will delete the attribue.

getAttrWords :: String -> Node -> [Text]

Remove an attribute

Get an attribute value, split into words

setAttrWords :: String -> [Text] -> Transform

Set an attribute with a list of words

addAttrWord

Arguments

:: String
attribute name

-> String
string to add (doesn't validate for wordiness)

-> Transform

Prepend a word to an attribute. (Won't create duplicates.)

delAttrWord

Arguments

:: String
attribute name

-> String
word to delete

-> Transform

Remove a word from anywhere in an attribute

addClass :: String -> Transform

Add a class name to an element. (Without duplicates.)

removeClass :: String -> Transform

Remove a class name from an element

replaceWith :: [Node] -> Transform

Replace the node with the provide node(s)

remove :: Transform

Remove the Node from the document

content :: [Node] -> Transform

Replace all of the node's children with the given node(s)

append :: [Node] -> Transform

Append node(s) after the existing children of this node. (i.e. just inside the closing tag of this node.)

prepend :: [Node] -> Transform

Prepend node(s) before existing children.

at :: [(Selector, Transform)] -> Transform

at allows you to nest selector-transformations. You can select a node in the usual fashion, then use at to select and transform only the children of that node. (This may be a premature optimisation. It may also be useful.)

Internal matching

checkPreds :: [Node -> Bool] -> Node -> Bool

Check a single selector step against a node.

data Match

data type to represent matches against node. PartialMatch represents a so-far-successful multi-step match.

Constructors

Match
Node matches Selector.

PartialMatch Selector
Node matches prefix of Selector. Contains remaining selector steps to be matched.

NoMatch
Node does not match.

matchSelector :: Selector -> Node -> Match

Match a single selector step, returning the appropriate Match.

High-level entry points

runTransforms :: [(Selector, Transform)] -> Document -> Document

run a list of `(Selector, Transform)` pairs across a document

transform :: Selector -> Transform -> Document -> Document

run a single Selector and Transform across a document

transform' :: Selector -> Transform -> [Node] -> [Node]

As above, but operates on [Node] rather than Document. Used internally, by at, and potentially for fragments that don't belong to a document.

nmap :: (Node -> [Node]) -> [Node] -> [[Node]]

A version of map for the list-tree-list structure of [Node].

docmap :: (Node -> [Node]) -> Document -> Document

Document-aware map, for the top-level.

modifyDocContent :: ([Node] -> [Node]) -> Document -> Document

Document-aware way to apply a function to all Nodes.

Examples

testDoc :: Document

A basic sample document

doTest :: Document

A (messy) example of some transforms

documentToString :: Document -> String

Utility function for debugging and testing.